lib/json: be more permissive with Attr inconsistencies
Before this patch, the json library would log.Fatal() when encoding a
HasAttrs struct whose .Attr(name) return nil or an error for a name that
was previously listed by AttrNames(). Fatal'ing in a library is a bit
presumptuous, particularly for a json encoding function that can be used
in misc scenarios where the output is meant for human consumption.  In
particular, Delve the Go debugger uses go-starlark to run scripts, and
also have implementations of HasAttr that can fail at runtime (e.g.  a
struct modelling a variable from the debugged process implements Attr
uses magic to present the field names of the target variable in
AttrNames, and the Attr implementation returns errors if one particular
field cannot be loaded).
Delve even goes as far as recovering panics from script execution, but
that doesn't work for log.Fatal.

This patch makes the json encoding function tolerant of Attr() either
returning an error, or otherwise returning a nil value. In both cases,
the output JSON will contain an error message as a string for the
respective field, and otherwise the encoding will continue.